data_processing/HaWoR/requirements.txt
numpy==1.26.4
opencv-python
pyrender
scikit-image
smplx==0.1.28
yacs
mmcv==1.3.9
timm
einops
xtcocotools
pandas
hydra-core
hydra-submitit-launcher
hydra-colorlog
pyrootutils
rich
webdataset
ultralytics
pulp
supervision
pycocotools
joblib
natsort
# git+https://github.com/facebookresearch/pytorch3d.git@stable
torch-scatter==2.1.2
evo
pytorch-minimize
mmengine==0.10.4
HTML4Vision
plyfile
aitviewer
easydict
loguru
dill
lapx
chumpy@git+https://github.com/mattloper/chumpy
moderngl-window==2.4.6